requirements.txt
jupyter==1.0.0
python-dotenv==1.0.0
openai==0.28.1
tiktoken==0.5.1
gensim==4.3.2
langchain==0.0.251
unstructured==0.7.9
faiss-cpu==1.7.4
altair==4.2.0
streamlit==1.16.0
streamlit-chat==0.1.1